feature: added custom hook for processing user info with use reducer
